widening_variable

+++ src/com/java/Type_casting_implicit.java
@@ -4,8 +4,8 @@ public class Type_casting_implicit {
 
 	public static void main(String[] args) {
      
-		int a=10;
-	    double b=a;
+		double a=10;
+	    int  b=(int)a;
 		
 		System.out.println(b);
